+++ Runtime/Behaviours/CinemachineClearShot.cs
@@ -29,11 +29,7 @@ namespace Cinemachine
     /// </summary>
     [DocumentationSorting(DocumentationSortingAttribute.Level.UserRef)]
     [DisallowMultipleComponent]
-#if UNITY_2018_3_OR_NEWER
     [ExecuteAlways]
-#else
-    [ExecuteInEditMode]
-#endif
     [ExcludeFromPreset]
     [AddComponentMenu("Cinemachine/CinemachineClearShot")]
     [HelpURL(Documentation.BaseURL + "manual/CinemachineClearShot.html")]