+++ CHANGELOG.md
@@ -4,8 +4,13 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [2.7.1-preview.1] - 2020-10-01
+## [2.7.0-preview.1] - 2020-10-10
 - Added ApplyAfter option to ImpulseListener, to add control over the ordering of extensions
+- Virtual Camera Lens inspector supports display of Horizontal FOV
+- Virtual Camera Lens can override orthographic and physical camera settings
+- Bugfix (1060230) - lens inspector sometimes displayed ortho vs perspective incorrectly for a brief time
+- Bugfix (1283984) - Error message when loading new scene with DontDestroyOnLoad
+- bugfix (1284701) - Edge-case exception when vcam is deleted
 
 
 ## [2.6.3] - 2020-09-16